Refactor: refactor-ft-backend-login
