+++ server/controllers/users/logout.js
